+++ BBDD/diseno/gold.py
@@ -1,7 +1,8 @@
-from datetime import datetime
 import conexionBBDD.conexionBBDD as conBBDD
 import pandas as pd
 import psycopg2
+import diseno.clases as clases
+CREATE_SQL_TABLES_FILE = "BBDD//diseno//gold_tables.sql"
 
 def diseno_gold(conn_str):
     df_video = pd.read_csv('video.csv')
@@ -10,156 +11,100 @@ def diseno_gold(conn_str):
     print("Leidos df_video y df_channel")
     conn=psycopg2.connect(conn_str)
     conn=conn.cursor()
-    #create_table_example(conn,"gold", "prueba", "employee_id SERIAL PRIMARY KEY,first_name VARCHAR(50),last_name VARCHAR(50),birth_date DATE,joined_date TIMESTAMP,department VARCHAR(50)")
     #drop_tables(conn,"gold")
-    create_tables(conn,"gold")
-    conn.connection.commit()
-    insert_data(conn, "gold",df_video, df_channel)
-    conn.connection.commit()
-    print("Datos insertados")
-    establish_relationships(conn,"gold")
-    conn.connection.commit()
-    print("Tablas conectadas")
+    #execute_sql_file(conn, CREATE_SQL_TABLES_FILE)
+    insert_data2(conn, "gold", df_channel, df_video)
 
-def create_table_example(conn, schema_name, table_name, attributes):
-    try:
-        # Create schema if it doesn't exist
-        conn.execute(f"CREATE SCHEMA IF NOT EXISTS {schema_name};")
-        
-        # Create table
-        query = f"""
-        CREATE TABLE IF NOT EXISTS {schema_name}.{table_name} (
-            {attributes}
-        );
+
+
+    
+def insert_data2(conn, schema_name, df_channel, df_video):
+    #insert_date(conn, schema_name, df_channel, "creationdate", "id")
+    #insert_date(conn, schema_name, df_video, "publishedAt", "id")
+    #insert_channel_statistics(conn, schema_name, df_channel)
+    insert_video_statistics(conn, schema_name, df_video)
+    #insert_video_and_tags(conn, schema_name, df_video)
+
+def insert_video_and_tags(conn, schema_name, df):
+    for index, row in df.iterrows():
+      print(index)
+      video = clases.create_video_from_csv(row)
+      try:
+        insert_video_query = f"""
+        INSERT INTO {schema_name}.video (video_id, title, statistics_id, category_id, duration, default_audio_language, type)
+        VALUES (%s, %s, %s, %s, %s, %s, %s);
         """
-        conn.execute(query)
-        print(f"Table {table_name} created successfully in schema {schema_name}")
+        conn.execute(insert_video_query, (video.video_id, video.title, video.statistics_id, video.category_id, video.duration, video.default_audio_language, video.type))
+        for tag in video.tags:
+          try:
+            insert_tag_query = f"""
+            INSERT INTO {schema_name}.tag (tag_id,tag_name, video_id)
+            VALUES (%s, %s);
+            """  
+            conn.execute(insert_tag_query, (tag+"_id", tag, video.video_id))
+          except Exception as e:
+            print(f"Error al insertar el tag {tag.tag_id}: {e}")
+            break
+
+      except Exception as e:
+        print(f"Error al insertar el video {video.video_id}: {e}")
+        break
+    conn.connection.commit()
+
+def insert_video_statistics(conn, schema_name, df):
+    for index, row in df.iterrows():
+      print(index)
+      video_stats = clases.create_video_statistics_from_csv(row)
+      try:
+        insert_channel_stats_query = f"""
+        INSERT INTO {schema_name}.videostatistics(stats_id, view_count, like_count, comment_count)
+        VALUES (%s, %s, %s, %s);
+        """
+        conn.execute(insert_channel_stats_query, (video_stats.stats_id,  video_stats.view_count, video_stats.like_count, video_stats.comment_count))
+      except Exception as e:
+        print(f"Error al insertar las estadísticas del canal {video_stats.stats_id}: {e}")
+        break
+    conn.connection.commit()
+
+def insert_channel_statistics(conn, schema_name, df):
+  REDONDEO = 1000
+  for index, row in df.iterrows():
+    print(index)
+    channel_stats = clases.create_channel_statistics_from_csv(row)
+    try:
+      insert_channel_stats_query = f"""
+      INSERT INTO {schema_name}.channelstatistics (stats_id, subscribers, view_count, num_videos)
+      VALUES (%s, %s, %s, %s);
+      """
+      conn.execute(insert_channel_stats_query, (channel_stats.stats_id,  channel_stats.subscribers, channel_stats.view_count/REDONDEO, channel_stats.num_videos))
     except Exception as e:
-        print(f"An error occurred: {e}")
+        print(f"Error al insertar las estadísticas del canal {channel_stats.stats_id}: {e}")
+        break
+  conn.connection.commit()
 
-def create_tables(conn, schema_name):
-  
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.dates (
-      date_id VARCHAR(255) PRIMARY KEY,
-      year INT NOT NULL,
-      month INT NOT NULL,
-      day INT NOT NULL,
-      hour INT NOT NULL,
-      min INT NOT NULL,
-      sec INT NOT NULL
-    )
-  """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'dates' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'dates': {e}")
+def insert_date(conn, schema_name, df, column_date, column_identifier):
+  for index, row in df.iterrows():
+    print(index)
+    date = clases.create_date_from_csv(row, df, column_date, column_identifier)
+    try:
+        insert_date_query = f"""
+        INSERT INTO {schema_name}.dates (date_id, year, month, day, hour, min, sec)
+        VALUES (%s, %s, %s, %s, %s, %s, %s);
+        """
+        conn.execute(insert_date_query, (date.date_id, date.year, date.month, date.day, date.hour, date.min, date.sec))
+    except Exception as e:
+        print(f"Error al insertar la fecha {date.date_id}: {e}")
+  conn.connection.commit()
 
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.Video (
-      video_id VARCHAR(255) PRIMARY KEY,
-      title VARCHAR(255) NOT NULL,
-      statistics_id VARCHAR(255) NOT NULL,
-      category_id INT NOT NULL,
-      duration VARCHAR(255) NOT NULL,
-      default_audio_language VARCHAR(255) NOT NULL,
-      type VARCHAR(255) NOT NULL
-    )
-  """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'Video' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'Video': {e}")
+     
 
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.Channel (
-        channel_id VARCHAR(255) PRIMARY KEY,
-        name VARCHAR(255) NOT NULL,
-        statistics_id VARCHAR(255) NOT NULL,
-        language VARCHAR(255) NOT NULL,
-        country VARCHAR(255) NOT NULL
-        )
-  """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'Channel' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'Channel': {e}")
 
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.VideoStatistics (
-        stats_id VARCHAR(255) PRIMARY KEY,
-        view_count INT NOT NULL,
-        like_count INT NOT NULL,
-        comment_count INT NOT NULL
-        )
-    """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'VideoStatistics' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'VideoStatistics': {e}")
-
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.ChannelStatistics (
-        stats_id VARCHAR(255) PRIMARY KEY,
-        subscribers INT NOT NULL,
-        view_count INT NOT NULL,
-        num_videos INT NOT NULL
-        )
-    """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'ChannelStatistics' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'ChannelStatistics': {e}")
-
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.VideoStats (
-        stats_id VARCHAR(255) PRIMARY KEY,
-        date_id VARCHAR(255) NOT NULL,
-        video_id VARCHAR(255) NOT NULL,
-        channel_id VARCHAR(255) NOT NULL
-        )
-  """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'VideoStats' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'VideoStats': {e}")
-
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.ChannelStats (
-        stats_id VARCHAR(255) PRIMARY KEY,
-        date_id VARCHAR(255) NOT NULL,
-        channel_id VARCHAR(255) NOT NULL
-        )
-  """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'ChannelStats' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'ChannelStats': {e}")
-
-  create_table_query = f"""
-    CREATE TABLE IF NOT EXISTS {schema_name}.Tag (
-        tag_id SERIAL PRIMARY KEY,
-        tag_name VARCHAR(255) NOT NULL
-        )
-    """
-  try:
-    conn.execute(create_table_query)
-    print("Tabla 'Tag' creada correctamente")
-  except Exception as e:
-    print(f"Error al crear la tabla 'Tag': {e}")
 
 def insert_data(conn, schema_name ,df_video, df_channel):
   
   for row in df_video.itertuples():
     try:
-      dates = create_date_from_csv(row, df_video)
+      dates = clases.create_date_from_csv(row, df_video)
     except UnicodeDecodeError:
        continue
     insert_date_query = f"""
@@ -172,7 +117,7 @@ def insert_data(conn, schema_name ,df_video, df_channel):
 
   for row in df_video.itertuples():
     try:
-      video = create_video_from_csv(row, df_video)
+      video = clases.create_video_from_csv(row, df_video)
     except UnicodeDecodeError:
        continue
     insert_video_query = f"""
@@ -184,7 +129,7 @@ def insert_data(conn, schema_name ,df_video, df_channel):
 
   for row in df_channel.itertuples():
     try:
-      channel = create_channel_from_csv(row, df_channel)
+      channel = clases.create_channel_from_csv(row, df_channel)
     except UnicodeDecodeError:
        continue
     insert_channel_query = f"""
@@ -195,7 +140,7 @@ def insert_data(conn, schema_name ,df_video, df_channel):
 
   for row in df_video.itertuples():
     try:
-      video_stats = create_video_stats_from_csv(row, video, channel, dates)
+      video_stats = clases.create_video_stats_from_csv(row, video, channel, dates)
     except UnicodeDecodeError:
        continue
     insert_video_stats_query = f"""
@@ -206,7 +151,7 @@ def insert_data(conn, schema_name ,df_video, df_channel):
 
   for row in df_channel.itertuples():
     try:
-      channel_stats = create_channel_stats_from_csv(row, channel, dates)
+      channel_stats = clases.create_channel_stats_from_csv(row, channel, dates)
     except UnicodeDecodeError:
        continue
     insert_channel_stats_query = f"""
@@ -215,183 +160,35 @@ def insert_data(conn, schema_name ,df_video, df_channel):
     """
     conn.execute(insert_channel_stats_query, (channel_stats.stats_id, channel_stats.dates.date_id, channel_stats.channel.channel_id))
 
-# Funciones para crear instancias a partir de los datos CSV
-def create_date_from_csv(row, df_video):
-    date_time_obj = datetime.strptime(row[4], '%Y-%m-%dT%H:%M:%SZ')
-    date_id = row[4] + "_" + row[3]
-    return dates(date_id, date_time_obj.year, date_time_obj.month, date_time_obj.day, date_time_obj.hour, date_time_obj.minute, date_time_obj.second)
-
-def create_video_from_csv(row):
-    tags = row['tags'].strip("[]").split(", ")
-    return Video(row['id'], row['title'], row['id'], row['categoryId'], row['duration'], row['defaultAudioLanguage'], row['type'], tags)
-
-def create_channel_from_csv(row):
-    return Channel(row['id'], row['name'], row['id'], row['language'], row['country'])
-
-def create_video_statistics_from_csv(row):
-    return VideoStatistics(row['id'], row['viewCount'], row['likeCount'], row['commentCount'])
-
-def create_channel_statistics_from_csv(row):
-    return ChannelStatistics(row['id'], row['subscribers'], row['viewcount'], row['numvideos'])
-
-def create_video_stats_from_csv(row, video, channel, dates):
-    return VideoStats(row['id'], dates, video, channel)
-
-def create_channel_stats_from_csv(row, channel, dates):
-    return ChannelStats(row['id'], dates, channel)
 
 
-def establish_relationships(conn, schema_name):
-
-    # Establecer la relación entre "VideoStats" y "Video"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.Video
-    ADD FOREIGN KEY (statistics_id) REFERENCES VideoStatistics(stats_id)
-    """
-    conn.execute(alter_table_query)
-
-    # Establecer la relación entre "ChannelStats" y "Channel"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.Channel
-    ADD FOREIGN KEY (statistics_id) REFERENCES ChannelStatistics(stats_id)
-    """
-    conn.execute(alter_table_query)
-
-    # Establecer la relación entre "VideoStats" y "Date"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.VideoStats
-    ADD FOREIGN KEY (date_id) REFERENCES dates(date_id)
-    """
-    conn.execute(alter_table_query)
-
-    # Establecer la relación entre "VideoStats" y "Video"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.VideoStats
-    ADD FOREIGN KEY (video_id) REFERENCES Video(video_id)
-    """
-    conn.execute(alter_table_query)
-
-    # Establecer la relación entre "VideoStats" y "Channel"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.VideoStats
-    ADD FOREIGN KEY (channel_id) REFERENCES Channel(channel_id)
-    """
-    conn.execute(alter_table_query)
-
-    # Establecer la relación entre "ChannelStats" y "Date"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.ChannelStats
-    ADD FOREIGN KEY (date_id) REFERENCES dates(date_id)
-    """
-    conn.execute(alter_table_query)
-
-    # Establecer la relación entre "ChannelStats" y "Channel"
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.ChannelStats
-    ADD FOREIGN KEY (channel_id) REFERENCES Channel(channel_id)
-    """
-    conn.execute(alter_table_query)
-
-    alter_table_query = f"""
-    ALTER TABLE {schema_name}.Tag
-    ADD FOREIGN KEY (video_id) REFERENCES Video(video_id)
-    """
-    conn.execute(alter_table_query)
-
-# Clases para representar las tablas de la base de datos
-
-class dates:
-    def __init__(self, date_id, year, month, day, hour, min, sec):
-        self.date_id = date_id
-        self.year = year
-        self.month = month
-        self.day = day
-        self.hour = hour
-        self.min = min
-        self.sec = sec
-
-class Video:
-    def __init__(self, video_id, title, statistics_id, category_id, duration, default_audio_language, type, tags=[]):
-        self.video_id = video_id
-        self.title = title
-        self.statistics_id = statistics_id
-        self.category_id = category_id
-        self.duration = duration
-        self.default_audio_language = default_audio_language
-        self.type = type
-        self.tags = tags
-
-class Channel:
-    def __init__(self, channel_id, name, statistics_id, language, country):
-        self.channel_id = channel_id
-        self.name = name
-        self.statistics_id = statistics_id
-        self.language = language
-        self.country = country
-    
-class VideoStatistics:
-    def __init__(self, stats_id, view_count, like_count, comment_count):
-        self.stats_id = stats_id
-        self.view_count = view_count
-        self.like_count = like_count
-        self.comment_count = comment_count
-
-class ChannelStatistics:
-    def __init__(self, stats_id, subscribers, view_count, num_videos):
-        self.stats_id = stats_id
-        self.subscribers = subscribers
-        self.view_count = view_count
-        self.num_videos = num_videos
-
-class VideoStats:
-    def __init__(self, stats_id, dates, video, channel):
-        self.stats_id = stats_id
-        self.dates = dates
-        self.video = video
-        self.channel = channel
-
-class ChannelStats:
-    def __init__(self, stats_id, dates, channel):
-        self.stats_id = stats_id
-        self.dates = dates
-        self.channel = channel
-
-    def get_stats_id(self):
-        return self.stats_id
-
-    def get_date(self):
-        return self.dates
-
-    def get_channel(self):
-        return self.channel
-
-    def set_stats_id(self, stats_id):
-        self.stats_id = stats_id
-
-    def set_date(self, dates):
-        self.dates = dates
-
-    def set_channel(self, channel):
-        self.channel = channel
+def execute_sql_file(cursor, sql_file_path):
+    with open(sql_file_path, 'r') as sql_file:
+        sql_commands = sql_file.read()
+    try:
+      cursor.execute(sql_commands)
+      cursor.connection.commit()
+      print(f"Fichero sql {sql_file_path} ejecutado correctamente")
+    except Exception as e:
+        print(f"Something went wrong: {e}") 
 
 
 def drop_tables(conn,schema_name):
-
         tables = [
-            "date",
             "dates",
-            "Video",
-            "Channel",
-            "VideoStatistics",
-            "ChannelStatistics",
-            "VideoStats",
-            "ChannelStats",
-            "Tag"
+            "video",
+            "channel",
+            "videostatistics",
+            "channelstatistics",
+            "videostats",
+            "channelstats",
+            "tag"
         ]
 
         for table in tables:
             try:
                 conn.execute(f"DROP TABLE IF EXISTS {schema_name}.{table} CASCADE;")
+                conn.connection.commit()
                 print(f"Tabla {table} eliminada correctamente")
             except psycopg2.Error as e:
                 print(f"Error al eliminar la tabla {table}: {e}")
