+++ BBDD/main.py
@@ -2,8 +2,6 @@ import preprocesado.bronze as bronze
 import preprocesado.silver as silver
 import diseno.gold as gold
 import conexionBBDD.conexionBBDD as conBBDD
-
-
 if __name__ == "__main__":
     conn_str = conBBDD.get_bbdd_url()
     test_conexion = conBBDD.verificar_conexion(conn_str)
@@ -12,3 +10,4 @@ if __name__ == "__main__":
     #bronze.create_bronze_schema(conn_str)
     #silver.preprocess_silver_scheme(conn_str)
     gold.diseno_gold(conn_str)
+
